QA-12314 chore: renaming and move package
